+++ 2-platforms/craft/craft/plugins/bust/variables/BustVariable.php
@@ -5,7 +5,7 @@ class BustVariable
 {
 	public function er($file)
 	{
-        $time = (file_exists($_SERVER['DOCUMENT_ROOT'] . $file)) ? filemtime($_SERVER['DOCUMENT_ROOT'] . $file) : '';
+		$time = filemtime($_SERVER['DOCUMENT_ROOT'] . $file);
 		return $file . '?' . $time;
 	}
 }