+++ 2-platforms/craft/craft/plugins/import/controllers/ImportController.php
@@ -20,7 +20,6 @@ class ImportController extends BaseController
      */
     public function actionGetEntryTypes()
     {
-
         // Only ajax post requests
         $this->requirePostRequest();
         $this->requireAjaxRequest();
@@ -41,7 +40,6 @@ class ImportController extends BaseController
      */
     public function actionUpload()
     {
-
         // Get import post
         $import = craft()->request->getRequiredPost('import');
 
@@ -51,38 +49,53 @@ class ImportController extends BaseController
         // Is file valid?
         if (!is_null($file)) {
 
-            // Determine folder
-            $folder = craft()->path->getStoragePath().'import/';
+            // Is asset source valid?
+            if (isset($import['assetsource']) && !empty($import['assetsource'])) {
 
-            // Ensure folder exists
-            IOHelper::ensureFolderExists($folder);
+                // Get source
+                $source = craft()->assetSources->getSourceTypeById($import['assetsource']);
 
-            // Get filepath - save in storage folder
-            $path = $folder.$file->getName();
+                // Get folder to save to
+                $folderId = craft()->assets->getRootFolderBySourceId($import['assetsource']);
 
-            // Save file to Craft's temp folder for later use
-            $file->saveAs($path);
+                // Save file to Craft's temp folder for later use
+                $fileName = AssetsHelper::cleanAssetName($file->name);
+                $filePath = AssetsHelper::getTempFilePath($file->extensionName);
+                $file->saveAs($filePath);
 
-            // Put vars in model
-            $model           = new ImportModel();
-            $model->filetype = $file->getType();
+                // Move the file by source type implementation
+                $response = $source->insertFileByPath($filePath, $folderId, $fileName, true);
 
-            // Validate filetype
-            if ($model->validate()) {
+                // Prevent sensitive information leak. Just in case.
+                $response->deleteDataItem('filePath');
 
-                // Get columns
-                $columns = craft()->import->columns($path);
+                // Get file id
+                $fileId = $response->getDataItem('fileId');
 
-                // Send variables to template and display
-                $this->renderTemplate('import/_map', array(
-                    'import'    => $import,
-                    'file'      => $path,
-                    'columns'   => $columns,
-                ));
+                // Put vars in model
+                $model = new ImportModel();
+                $model->filetype = $file->getType();
+
+                // Validate filetype
+                if ($model->validate()) {
+
+                    // Get columns
+                    $columns = craft()->import->columns($fileId);
+
+                    // Send variables to template and display
+                    $this->renderTemplate('import/_map', array(
+                        'import' => $import,
+                        'file' => $fileId,
+                        'columns' => $columns,
+                    ));
+                } else {
+
+                    // Not validated, show error
+                    craft()->userSession->setError(Craft::t('This filetype is not valid').': '.$model->filetype);
+                }
             } else {
-
-                // Not validated, show error
-                craft()->userSession->setError(Craft::t('This filetype is not valid').': '.$model->filetype);
+                // No asset source selected
+                craft()->userSession->setError(Craft::t('Please select an asset source.'));
             }
         } else {
 
@@ -96,29 +109,30 @@ class ImportController extends BaseController
      */
     public function actionImport()
     {
-
         // Get import post
         $settings = craft()->request->getRequiredPost('import');
 
         // Get file
-        $file = craft()->request->getParam('file');
+        $fileId = craft()->request->getParam('file');
+        $file = craft()->assets->getFileById($fileId);
 
         // Get mapping fields
         $map = craft()->request->getParam('fields');
         $unique = craft()->request->getParam('unique');
 
         // Get rows/steps from file
-        $rows = count(craft()->import->data($file));
+        $rows = count(craft()->import->data($file->id));
 
         // Proceed when atleast one row
         if ($rows) {
 
             // Set more settings
             $settings = array_merge(array(
-                'file'        => $file,
-                'rows'        => $rows,
-                'map'         => $map,
-                'unique'      => $unique,
+                'user' => craft()->userSession->getUser()->id,
+                'file' => $file->id,
+                'rows' => $rows,
+                'map' => $map,
+                'unique' => $unique,
             ), $settings);
 
             // Create history
@@ -130,18 +144,8 @@ class ImportController extends BaseController
             // UNCOMMENT FOR DEBUGGING
             //craft()->import->debug($settings, $history, 1);
 
-            // Determine new folder to save original importfile
-            $folder = dirname($file).'/'.$history.'/';
-            IOHelper::ensureFolderExists($folder);
-
-            // Move the file to its history folder
-            IOHelper::move($file, $folder.basename($file));
-
-            // Update the settings with the new file location
-            $settings['file'] = $folder.basename($file);
-
             // Create the import task
-            $task = craft()->tasks->createTask('Import', Craft::t('Importing').' '.basename($file), $settings);
+            $task = craft()->tasks->createTask('Import', Craft::t('Importing').' '.$file->filename, $settings);
 
             // Notify user
             craft()->userSession->setNotice(Craft::t('Import process started.'));