+++ 2-platforms/craft/craft/plugins/sproutforms/services/SproutForms_FormsService.php
@@ -81,7 +81,7 @@ class SproutForms_FormsService extends BaseApplicationComponent
 		$formRecord->groupId                  = $form->groupId;
 		$formRecord->redirectUri              = $form->redirectUri;
 		$formRecord->submitAction             = $form->submitAction;
-		$formRecord->savePayload              = $form->savePayload;
+		$formRecord->saveData                 = $form->saveData;
 		$formRecord->submitButtonText         = $form->submitButtonText;
 		$formRecord->notificationEnabled      = $form->notificationEnabled;
 		$formRecord->notificationRecipients   = $form->notificationRecipients;
@@ -631,9 +631,18 @@ class SproutForms_FormsService extends BaseApplicationComponent
 
 	public function createNewForm($name = null, $handle = null)
 	{
-		$form   = new SproutForms_FormModel();
-		$name   = empty($name) ? 'Form' : $name ;
-		$handle = empty($handle) ? 'form' : $handle;
+		$form     = new SproutForms_FormModel();
+		$name     = empty($name) ? 'Form' : $name ;
+		$handle   = empty($handle) ? 'form' : $handle;
+		$settings = craft()->plugins->getPlugin('sproutforms')->getSettings();
+
+		if ($settings['enableSaveData'])
+		{
+			if ($settings['enableSaveDataPerFormBasis'])
+			{
+				$form->saveData = $settings['saveDataByDefault'];
+			}
+		}
 
 		$form->name   = sproutForms()->forms->getFieldAsNew('name', $name);
 		$form->handle = sproutForms()->forms->getFieldAsNew('handle', $handle);
@@ -653,5 +662,17 @@ class SproutForms_FormsService extends BaseApplicationComponent
 		}
 
 		return false;
+	}
+
+	public function installDefaultSettings()
+	{
+		$defaultSettings = '{"pluginNameOverride":null,"templateFolderOverride":"","enablePerFormTemplateFolderOverride":"","enablePayloadForwarding":"","enableSaveData":"1","enableSaveDataPerFormBasis":"", "saveDataByDefault":"1"}';
+
+		craft()->db->createCommand()->update('plugins',
+				array('settings' => $defaultSettings),
+				array('class' => 'SproutForms')
+			);
+
+		return true;
 	}
 }